ai-suggestion-server/requirements.txt
annotated-types==0.7.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyttsx3==2.98
requests==2.32.3
rsa==4.9
SpeechRecognition==3.14.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
beautifulsoup4
gunicorn==23.0.0
Flask==3.0.3
flask-cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.4
click==8.1.7
colorama==0.4.6
zipp==3.20.0
six==1.16.0
certipy==0.1.3
colorlog==6.7.0
Flask-HTTPAuth==4.7.0
bcrypt==4.0.1
email-validator==2.0.0
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.0
alembic==1.11.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
